+++ CHANGELOG.md
@@ -79,6 +79,7 @@
 - Script Integration\Add SKIP tag at current playback: added warnings when trying to add SKIP tag to the end of start of a track.
 - Buttons: new separator element which may be added (multiple times) to the toolbar.
 - Buttons: new 'display_volume' button which may be used to display and control foobar2000 volume.
+- Buttons: new 'display_tf' button which may be used to display arbitrary Title Format expressions on the toolbar (evaluated on selection, now playing or globally). It also works with dynamic queries variables (see below).
 - Buttons: 'utils_main_menu' button now has additional settings to save and handle button states, either internally or by main menu checks.
 - Buttons: added more presets to remove/show duplicates and filter playlist buttons.
 - Readmes: added readme for 'buttons_device_switcher.js' button.